Add integration file. Generalize test helpers. Clean up Array<> usage to more TypeScript-esque []
