+++ src/testHelper.js
@@ -1,0 +1,8 @@
+// @flow
+export const rootStyle = { display: 'flex', justifyContent: 'center' };
+export const rowStyle = {
+  margin: '200px 0',
+  display: 'flex',
+  justifyContent: 'space-between',
+};
+export const boxStyle = { padding: '10px', border: '1px solid black' };
